Make xlC happy
git-svn-id: https://svn.smr.arpa.emr.it/dballe/wreport/trunk@3517 5c06e07b-eeef-0310-a3fe-9f68d50baefc

+++ wreport/compat.h
@@ -87,7 +87,7 @@ static int vasprintf (char **result, const char *format, va_list args)
 #endif
 
 #if USE_OWN_EXP10
-#define exp10(x) pow(10.0, (x))
+#define exp10(x) pow((double)10.0, (double)(x))
 #endif
 
 #endif
